// Server/src/Entity/Blueprint/enum.ts
export enum BlueprintId {
  Smelting_CopperIngot = "smeltingCopperIngot",
  Smelting_TinIngot = "smeltingTinIngot",
  Smelting_IronIngot = "smeltingIronIngot",
  Smelting_SilverIngot = "smeltingSilverIngot",
  Smelting_GoldIngot = "smeltingGoldIngot",
  Alloy_BronzeIngot = "alloyBronzeIngot",
  Alloy_SteelIngot = "alloySteelIngot",
  Alloy_ElectrumIngot = "alloyElectrumIngot",
  Infusion_AethersteelIngot = "infusionAethersteelIngot",
  Infusion_VoidforgedIngot = "infusionVoidforgedIngot",
  Weapon_Axe = "weaponAxe",
  Weapon_BroadAxe = "weaponBroadAxe",
  Weapon_WarAxe = "weaponWarAxe",
  Weapon_Knife = "weaponKnife",
  Weapon_Stiletto = "weaponStiletto",
  Weapon_ShortSword = "weaponShortSword",
  Weapon_LongSword = "weaponLongSword",
  Weapon_Rapier = "weaponRapier",
  Weapon_GreatSword = "weaponGreatSword",
  Weapon_Katana = "weaponKatana",
  Weapon_Scimitar = "weaponScimitar",
  Weapon_Cutlass = "weaponCutlass",
  Weapon_Falchion = "weaponFalchion",
  Weapon_Hammer = "weaponHammer",
  Weapon_MorningStar = "weaponMorningStar",
  Weapon_Scepter = "weaponScepter",
  Weapon_WarHammer = "weaponWarHammer",
  Weapon_Dory = "weaponDory",
  Weapon_Javelin = "weaponJavelin",
  Weapon_Halberd = "weaponHalberd",
  Weapon_ShortBow = "weaponShortBow",
  Weapon_LongBow = "weaponLongBow",
  Weapon_Crossbow = "weaponCrossbow",
  Weapon_Staff = "weaponStaff",
  Weapon_QuarterStaff = "weaponQuarterStaff",
  Weapon_LongestStaff = "weaponLongestStaff",
  Weapon_Buckler = "weaponBuckler",
  Weapon_KiteShield = "weaponKiteShield",
  Weapon_TowerShield = "weaponTowerShield",
  Weapon_Bible = "weaponBible",
  Weapon_Codex = "weaponCodex",
  Weapon_Grimoire = "weaponGrimoire",
  Weapon_Wand = "weaponWand",
  Weapon_Orb = "weaponOrb",
}